several minor updates  * move data fixtures to conftest  * added TODO in readme  * "maximized" fixture moved to test scope
