+++ tests/scenarios/readme.md
@@ -11,4 +11,9 @@ Tests in the scenarios directory will not be discovered by the py.test collectio
 
 ### Data
 
-The data subdirectory in each scenario is where the corresponding data files are placed. In it you will need to rename and customize the `cfme_data.yaml.template` file.
+The data subdirectory in each scenario is where the corresponding data files are placed. In it you will need to rename and customize the `cfme_data.yaml.template` file. The tests shall be generic and be driven by scenario data fixtures (`conftest.py`) plus the scenario data directory.
+
+### TODO
+
+ * Move more of the test workflows into service methods in cfme_pages repo so the tests are very simple. This will make test duplication more tolerable.
+ * Better still, consider ways to order test collection via py.test hooks. This would allow tests to move to a higher level directory and become a library that multiple scenarios could access.
